Refactor app folder
